Fixed size_t issue

+++ sph/include/sph/particles_data.hpp
@@ -77,8 +77,8 @@ public:
     ParticlesData()                     = default;
     ParticlesData(const ParticlesData&) = delete;
 
-    size_t iteration{1};
-    size_t numParticlesGlobal;
+    uint64_t iteration{1};
+    uint64_t numParticlesGlobal;
 
     //! @brief default mean desired number of neighbors per particle, can be overriden per test case or input file
     unsigned ng0{100};
@@ -147,7 +147,7 @@ public:
     }
 
     //! @brief non-stateful variables for statistics
-    size_t totalNeighbors;
+    uint64_t totalNeighbors;
 
     /*! @brief Particle fields
      *
